+++ DinnerApp/ui/favoriteStores/FavoriteStores.swift
@@ -1,0 +1,20 @@
+//
+//  FavoriteStores.swift
+//  DinnerApp
+//
+//  Created by hiroki.horiguchi on 2023/05/12.
+//
+
+import SwiftUI
+
+struct FavoriteStores: View {
+    var body: some View {
+        Text("お気に入り店舗一覧画面だよ。DB に追加したお気に入り店舗を取得してきて表示するよ")
+    }
+}
+
+struct FavoriteStores_Previews: PreviewProvider {
+    static var previews: some View {
+        FavoriteStores()
+    }
+}
