+++ DinnerApp/ui/storeIndex/StoreIndex.swift
@@ -1,0 +1,20 @@
+//
+//  StoreIndex.swift
+//  DinnerApp
+//
+//  Created by hiroki.horiguchi on 2023/05/12.
+//
+
+import SwiftUI
+
+struct StoreIndex: View {
+    var body: some View {
+        Text("店舗一覧画面だよ。ここに pager みたいなやつぶち込んで、BCE の foryou みたいにするよ")
+    }
+}
+
+struct SwiftUIView_Previews: PreviewProvider {
+    static var previews: some View {
+        StoreIndex()
+    }
+}
